controllerUtilities updates along with comment fixes.
