--- Fox_app/backend/controllers/Controller_scripts.js
@@ -1,4 +1,0 @@
-//uuidRegex to validate UUID format
-const uuidRegex = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
-
-module.exports = { uuidRegex };
